[수정] 평균 평점이 소수점으로 전달되는 경우 JSonDataException

+++ domain/src/main/java/com/beok/domain/entity/VolumeInfo.kt
@@ -55,7 +55,7 @@ data class VolumeInfo(
 	val description: String = "",
 
 	@Json(name = "averageRating")
-	val averageRating: Int = 0,
+	val averageRating: Double = 0.0,
 
 	@Json(name = "panelizationSummary")
 	val panelizationSummary: PanelizationSummary = PanelizationSummary(),
